project_mercury/rtm/src/main/java/com/smartsparrow/rtm/subscription/courseware/themechange/ActivityThemeChangeRTMConsumable.java
package com.smartsparrow.rtm.subscription.courseware.themechange;

import com.smartsparrow.pubsub.data.RTMEvent;
import com.smartsparrow.rtm.message.RTMClientContext;
import com.smartsparrow.rtm.subscription.courseware.ActivityRTMSubscription;
import com.smartsparrow.rtm.subscription.courseware.message.ConfigChangeBroadcastMessage;
import com.smartsparrow.rtm.subscription.data.AbstractRTMConsumable;

/**
 * This RTM consumable describes an activity theme change event
 */
public class ActivityThemeChangeRTMConsumable extends AbstractRTMConsumable<ConfigChangeBroadcastMessage> {

    private static final long serialVersionUID = -6918639607354911278L;

    public ActivityThemeChangeRTMConsumable(RTMClientContext rtmClientContext, ConfigChangeBroadcastMessage content) {
        super(rtmClientContext, content);
    }

    @Override
    public RTMClientContext getRTMClientContext() {
        return rtmClientContext;
    }

    @Override
    public String getName() {
        return String.format("author.activity/%s/%s", content.getActivityId(), getRTMEvent().getName());
    }

    @Override
    public String getSubscriptionName() {
        return ActivityRTMSubscription.NAME(content.getActivityId());
    }

    @Override
    public RTMEvent getRTMEvent() {
        return new ActivityThemeChangeRTMEvent();
    }

    @Override
    public boolean equals(Object o) {
        return super.equals(o);
    }

    @Override
    public int hashCode() {
        return super.hashCode();
    }
}
